# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CPay API
  version: v1
servers:
  - url: http://localhost:8080
paths:
...
  /v1/products:
    post:
      summary: Create product
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Created
    get:
      summary: List products
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Success
  /v1/products/{id}:
    get:
      summary: Get public product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
    post:
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Success
    delete:
      summary: Delete product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Deleted
  /v1/public/products/{id}:
    get:
      summary: Get public product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
        '404':
          description: Product not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key